Fix: refactor handleSubmit function on Form component with a custom hook
